[UPDATE] final SaleInvoice

+++ src/main/java/com/rmit/trading_backend/repository/ProviderRepository.java
@@ -5,8 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProviderRepository extends JpaRepository<Provider, Integer> {
+public interface ProviderRepository extends JpaRepository<Provider, Long> {
 
 
     Provider findProviderByEmail(String email);
+
 }
